requirements.txt
fastapi
uvicorn
python-multipart
replicate
requests
pytest
httpx
pillow
